move exception handling to match rest of team
